Move things around (Flatten project structure)
